+++ apis/vara-eth/test/setup/setup.ts
@@ -43,10 +43,6 @@ async function setupVaraEth() {
     {
       stdio: 'pipe',
       detached: true,
-      env: {
-        ...process.env,
-        RUST_LOG: 'debug',
-      },
     },
   );